ENH: include exon phase in cds spans
[NEW] now capture exon phase and end_phase in the exon_view table,
     these are used to adjust the rel_start and rel_end values to
     the limiting exons to get the correct cds spans

[CHANGED] modified tests to include these new columns

Notes: it seems rel_start, which derives from translation.seq_start
is not always correct for starting exons with rank > 1. It may be that
when rel_start is 0 the phase can be > 0, at which point we need to
compensate. If I can confirmt this is correct, the LimitingExon code
and attributes can be simplified.